fix: レビュー指摘事項の修正（import パス・zod バージョン・lint regex）
- brain-manager.test.ts: 相対パスを @vicissitude/store/test-helpers に統一
- scheduling: zod を ^3.24.4 → ^4.3.6 に統一（他パッケージと一致）
- .oxlintrc.json: ollama regex に境界指定 (?:/|$) を追加
- test-helpers を packages/shared に移動し spec/ から re-export
