+++ .oxlintrc.json
@@ -188,7 +188,7 @@
 					{
 						"patterns": [
 							{
-								"regex": "@vicissitude/(?!shared/|ollama)",
+								"regex": "@vicissitude/(?!shared/|ollama(?:/|$))",
 								"message": "ltm は shared, ollama のみに依存できます"
 							}
 						]
@@ -236,7 +236,7 @@
 					{
 						"patterns": [
 							{
-								"regex": "@vicissitude/(?!shared/|infrastructure/|ltm/|ltm$|ollama|store/)",
+								"regex": "@vicissitude/(?!shared/|infrastructure/|ltm/|ltm$|ollama(?:/|$)|store/)",
 								"message": "mcp は shared, infrastructure, ltm, ollama, store のみに依存できます"
 							}
 						]
