+++ packages/agent/src/minecraft/brain-manager.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 
 import { clearSessionLock, tryAcquireSessionLock } from "@vicissitude/store/mc-bridge";
+import { createTestDb } from "@vicissitude/store/test-helpers";
 
-import { createTestDb } from "../../../store/src/test-helpers.ts";
 import type { McBrainManagerDeps } from "./brain-manager.ts";
 import { McBrainManager } from "./brain-manager.ts";
