+++ packages/scheduling/package.json
@@ -7,6 +7,6 @@
 		"./consolidation-scheduler": "./src/consolidation-scheduler.ts"
 	},
 	"dependencies": {
-		"zod": "^3.24.4"
+		"zod": "^4.3.6"
 	}
 }
