+++ packages/shared/src/test-helpers.ts
@@ -1,0 +1,22 @@
+import { mock } from "bun:test";
+
+import type { Logger, MetricsCollector } from "./types";
+
+export function createMockLogger(): Logger {
+	return {
+		info: mock(() => {}),
+		error: mock(() => {}),
+		warn: mock(() => {}),
+	};
+}
+
+export function createMockMetrics(): MetricsCollector {
+	return {
+		incrementCounter: mock(() => {}),
+		addCounter: mock(() => {}),
+		setGauge: mock(() => {}),
+		incrementGauge: mock(() => {}),
+		decrementGauge: mock(() => {}),
+		observeHistogram: mock(() => {}),
+	};
+}
